requirements.txt
discord-py-interactions
bs4
requests